refactor(rpc): split resource types into per-resource files, remove app re-exports
A single resource-types.ts won't scale as we add types for the full v2
API. Per-resource directories (resources/pipeline/, resources/trigger/,
etc.) each with a types.ts give a natural home for future growth.

Consumers now import directly from @michelangelo-ai/rpc/resources/*
instead of through app re-export shims that added indirection without
value.
