--- javascript/app/config/entities/pipeline/types.ts
@@ -1,1 +1,0 @@
-export type { Pipeline } from '@michelangelo-ai/rpc/resource-types';
