--- javascript/app/config/entities/run/types.ts
@@ -1,1 +1,0 @@
-export type { PipelineRun } from '@michelangelo-ai/rpc/resource-types';
