--- javascript/app/config/entities/trigger/types.ts
@@ -1,2 +1,0 @@
-export { TriggerRunAction, TriggerRunState } from '@michelangelo-ai/rpc/resource-types';
-export type { Trigger, TriggerRun } from '@michelangelo-ai/rpc/resource-types';
