+++ javascript/packages/rpc/package.json
@@ -23,8 +23,14 @@
     ".": {
       "default": "./index.ts"
     },
-    "./resource-types": {
-      "default": "./resource-types.ts"
+    "./resources/pipeline": {
+      "default": "./resources/pipeline/types.ts"
+    },
+    "./resources/pipeline-run": {
+      "default": "./resources/pipeline-run/types.ts"
+    },
+    "./resources/trigger": {
+      "default": "./resources/trigger/types.ts"
     }
   },
   "imports": {
