+++ javascript/packages/rpc/resources/pipeline-run/types.ts
@@ -1,0 +1,18 @@
+export type PipelineRun = {
+  metadata: {
+    name: string;
+    namespace: string;
+  };
+  spec: {
+    /** Populated server-side from the `x-user-name` request header, not set by the client. */
+    actor?: {
+      name: string;
+    };
+    pipeline: {
+      name: string;
+      namespace: string;
+    };
+    /** Optional human-readable description for this run. */
+    description?: string;
+  };
+};
