+++ javascript/packages/rpc/resources/pipeline/types.ts
@@ -1,0 +1,11 @@
+export interface Pipeline {
+  metadata: {
+    name: string;
+    namespace: string;
+  };
+  spec: {
+    owner: {
+      name: string;
+    };
+  };
+}
