+++ javascript/packages/rpc/resources/trigger/types.ts
@@ -1,45 +1,3 @@
-/**
- * Curated API model types — hand-written simplified views of proto-generated
- * shapes that include only the fields the UI needs. Update alongside the
- * corresponding proto definitions.
- */
-
-export interface Pipeline {
-  metadata: {
-    name: string;
-    namespace: string;
-  };
-  spec: {
-    owner: {
-      name: string;
-    };
-  };
-}
-
-export type PipelineRun = {
-  metadata: {
-    name: string;
-    namespace: string;
-  };
-  spec: {
-    /** Populated server-side from the `x-user-name` request header, not set by the client. */
-    actor?: {
-      name: string;
-    };
-    pipeline: {
-      name: string;
-      namespace: string;
-    };
-    /** Optional human-readable description for this run. */
-    description?: string;
-  };
-};
-
-/**
- * Mirrors generated types from trigger_run_pb.
- * Update alongside proto/api/v2/trigger_run.proto.
- */
-
 export type Trigger = {
   metadata: {
     name: string;
